src/requirements.txt
flask==3.0.0
alembic==1.13.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
psycopg2==2.9.9